+++ FatanAbdullahLabe22/FatanAbdullahLabe22.xcodeproj/xcuserdata/faten.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "7CAFA0F9-75C3-48C2-8567-1DBC61DC722D"
+   type = "1"
+   version = "2.0">
+</Bucket>
