+++ FatanAbdullahLabe22/FatanAbdullahLabe22.xcodeproj/xcuserdata/faten.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -4,7 +4,7 @@
 <dict>
 	<key>SchemeUserState</key>
 	<dict>
-		<key>FatanLabe22.xcscheme_^#shared#^_</key>
+		<key>FatanAbdullahLabe22.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
 			<integer>0</integer>
