+++ FatanAbdullahLabe22/FatanAbdullahLabe22/Model/Digimon.swift
@@ -1,6 +1,6 @@
 //
 //  Digimon.swift
-//  FatanLabe22
+//  FatanAbdullahLabe22
 //
 //  Created by Faten Abdullh salem on 22/04/1443 AH.
 //
@@ -11,3 +11,6 @@ var img:String
 var name:String
 var level:String
 }
+    
+
+
