+++ FatanAbdullahLabe22/FatanAbdullahLabe22/View/digimonDetailCell.swift
@@ -1,0 +1,15 @@
+//
+//  digimonDetailCell.swift
+//  FatanAbdullahLabe22
+//
+//  Created by Faten Abdullh salem on 22/04/1443 AH.
+//
+
+import Foundation
+import UIKit
+class digimonDetailCell:UITableViewCell {
+@IBOutlet weak var digimonImage: UIImageView!
+@IBOutlet weak var digimonName: UILabel!
+@IBOutlet weak var digimonLevel: UILabel!
+    
+}
